+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 
 - The Employee Card won't spawn for Sonic if the level is not accessible for him.
 - Fixed Sonic's Speed Highway Entrance not needing the in-game Employee Card.
+- Speed Highway's shutter will open as soon as you get the Employee Card for Amy and Gamma.
+    - Previously, you had to leave station square if you got card in the main area.
 
 ## [v0.8.5] - 2024-09-08